Reorganize includes for sound headers
