+++ headers/types/types.h
@@ -2,7 +2,6 @@
 #define HEADERS_TYPES_H_
 
 #include "common/common.h"
-#include "common/sound.h"
 #include "dungeon_mode/dungeon_mode.h"
 #include "ground_mode/ground_mode.h"
 #include "sentry_duty/sentry_duty.h"
